refactor(host-application): dedupe hook security helpers
Extract shared hook trust and parsed hook execution helpers into app_service/hook_security and reuse in build/runtime orchestrators.
